rename top level expression module to eval
