+++ fluree-db-query/src/eval/arithmetic.rs
