+++ fluree-db-query/src/eval/compare.rs
