+++ fluree-db-query/src/eval/fulltext.rs
