+++ fluree-db-query/src/eval/geo.rs
