+++ fluree-db-query/src/eval/logical.rs
