+++ fluree-db-query/src/eval/numeric.rs
