+++ fluree-db-query/src/eval/string.rs
