+++ fluree-db-query/src/eval/vector.rs
