+++ fluree-db-query/src/eval/vector_math.rs
