+++ fluree-db-query/tests/correctness_tests.rs
@@ -10,7 +10,7 @@ use fluree_db_core::{LedgerSnapshot, Sid};
 use fluree_db_query::binary_scan::EmitMask;
 use fluree_db_query::binding::{Batch, Binding};
 use fluree_db_query::context::ExecutionContext;
-use fluree_db_query::expression::PreparedBoolExpression;
+use fluree_db_query::eval::PreparedBoolExpression;
 use fluree_db_query::ir::triple::{Ref, Term, TriplePattern};
 use fluree_db_query::ir::{Expression, FlakeValue};
 use fluree_db_query::join::NestedLoopJoinOperator;
